Require file extensions to be specified in directives.
`mountain join` used to automatically try a list of known file
extensions, in the event that the user did not provide a file extension
when referencing a file in an #include directive.

Thus, you could write something like this:

    [[#include intro]]

And that file would automatically be found, provided that a file named
`intro` existed with a `.fountain`, `.txt`, `.md`, `.mdown`, or
`.markdown`extension.

This commit drops support for this magic. Instead, you'll need to write:

    [[#include intro.fountain]]

(Or use whatever file extension you prefer.)

Although the code was there, and it worked, it required some magic that
may be confusing to explain to users and I'm generally worried that it
will introduce more issues than it solves.

It wasn't yet implemented in the reverse case (`mountain split`), but
implementation is feasible. Here's the algorithm:

- check if file_path ends with an extension
    - (if yes, write the file; return)
- check if file_path exists
    - (if yes, write the file; return)
- otherwise, file_path does not exist
    - for each supported extension, add the extension to file_path and check if
      the file_path exists
      - if it does, write the file; break
      - else, create a new file with the default extension (.fountain)